# tequilapi/docs/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "The purpose of this documentation is to provide developers an insight of how to\ninteract with Mysterium Node via Tequila API.\nThis should demonstrate all the possible API calls with described parameters and responses.",
    "title": "Tequila API",
    "version": "dev"
  },
  "host": "127.0.0.1:4050",
  "paths": {
...
    "/identities": {
      "get": {
        "description": "Returns list of identities",
        "tags": [
          "Identity"
        ],
        "summary": "Returns identities",
        "operationId": "listIdentities",
        "responses": {
          "200": {
            "description": "List of identities",
            "schema": {
              "$ref": "#/definitions/ListIdentitiesResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorMessageDTO"
            }
          }
        }
      },
      "post": {
        "description": "Creates identity and stores in keystore encrypted with passphrase",
        "tags": [
          "Identity"
        ],
        "summary": "Creates new identity",
        "operationId": "createIdentity",
        "parameters": [
          {
            "description": "Parameter in body (passphrase) required for creating new identity",
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/IdentityCreateRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Identity created",
            "schema": {
              "$ref": "#/definitions/IdentityRefDTO"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorMessageDTO"
            }
          },
          "422": {
            "description": "Parameters validation error",
            "schema": {
              "$ref": "#/definitions/ValidationErrorDTO"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorMessageDTO"
            }
          }
        }
      }
    },
    "/identities-import": {
      "post": {
        "description": "Imports a given identity returning it is a blob of text which can later be used to import it back.",
        "tags": [
          "Identities"
        ],
        "summary": "Imports a given identity.",
        "operationId": "importIdentity",
        "parameters": [
          {
            "description": "Parameter in body used to import an identity.",
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/IdentityImportRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unlocked identity returned",
            "schema": {
              "$ref": "#/definitions/IdentityRefDTO"
            }
          },
          "400": {
            "description": "Bad Request error",
            "schema": {
              "$ref": "#/definitions/ErrorMessageDTO"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorMessageDTO"
            }
          }
        }
      }
    },
    "/identities/current": {
      "put": {
        "description": "Tries to retrieve the last used identity, the first identity, or creates and returns a new identity",
        "tags": [
          "Identity"
        ],
        "summary": "Returns my current identity",
        "operationId": "currentIdentity",
        "parameters": [
          {
            "description": "Parameter in body (passphrase) required for creating new identity",
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/IdentityCurrentRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unlocked identity returned",
            "schema": {
              "$ref": "#/definitions/IdentityRefDTO"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorMessageDTO"
            }
          },
          "422": {
            "description": "Parameters validation error",
            "schema": {
              "$ref": "#/definitions/ValidationErrorDTO"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorMessageDTO"
            }
          }
        }
      }
    },
...
    "/identities/{id}": {
      "get": {
        "description": "Provide identity details",
        "tags": [
          "Identity"
        ],
        "summary": "Get identity",
        "operationId": "getIdentity",
        "parameters": [
          {
            "type": "string",
            "description": "hex address of identity",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Identity retrieved",
            "schema": {
              "$ref": "#/definitions/IdentityRefDTO"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorMessageDTO"
            }
          }
        }
      }
    },
    "/identities/{id}/balance/refresh": {
      "put": {
        "description": "Refresh balance of given identity",
        "tags": [
          "Identity"
        ],
        "summary": "Refresh balance of given identity",
        "operationId": "balance",
        "parameters": [
          {
            "type": "string",
            "description": "hex address of identity",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Updated balance",
            "schema": {
              "$ref": "#/definitions/BalanceDTO"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorMessageDTO"
            }
          }
        }
      }
    },
    "/identities/{id}/beneficiary": {
      "get": {
        "description": "Provides beneficiary address for given identity",
        "tags": [
          "Identity",
          "beneficiary"
        ],
        "summary": "Provide identity beneficiary address",
        "operationId": "address",
        "parameters": [
          {
            "type": "string",
            "description": "hex address of identity",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Beneficiary retrieved",
            "schema": {
              "$ref": "#/definitions/IdentityBeneficiaryResponseDTO"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorMessageDTO"
            }
          }
        }
      }
    },
...
    "/identities/{id}/referral": {
      "get": {
        "description": "Gets a referral token for the given identity if a campaign exists",
        "summary": "Gets referral token",
        "operationId": "Referral",
        "parameters": [
          {
            "type": "string",
            "description": "Identity for which to get a token",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Token response"
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorMessageDTO"
            }
          }
        }
      }
    },
    "/identities/{id}/referral-available": {
      "get": {
        "description": "Verifies user's eligibility and the presence of an applicable public campaign",
        "tags": [
          "Referral",
          "availability"
        ],
        "summary": "Checks if the user can obtain a referral token",
        "operationId": "check",
        "parameters": [
          {
            "type": "string",
            "description": "Identity for which to get a token",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorMessageDTO"
            }
          }
        }
      }
    },
...
    "/identities/{id}/registration": {
      "get": {
        "description": "Provides registration status for given identity, if identity is not registered - provides additional data required for identity registration",
        "tags": [
          "Identity"
        ],
        "summary": "Provide identity registration status",
        "operationId": "identityRegistration",
        "parameters": [
          {
            "type": "string",
            "description": "hex address of identity",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Status retrieved",
            "schema": {
              "$ref": "#/definitions/IdentityRegistrationResponseDTO"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorMessageDTO"
            }
          }
        }
      }
    },
    "/identities/{id}/unlock": {
      "put": {
        "description": "Uses passphrase to decrypt identity stored in keystore",
        "tags": [
          "Identity"
        ],
        "summary": "Unlocks identity",
        "operationId": "unlockIdentity",
        "parameters": [
          {
            "type": "string",
            "description": "Identity stored in keystore",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Parameter in body (passphrase) required for unlocking identity",
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/IdentityUnlockRequestDTO"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Identity unlocked"
          },
          "400": {
            "description": "Body parsing error",
            "schema": {
              "$ref": "#/definitions/ErrorMessageDTO"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorMessageDTO"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorMessageDTO"
            }
          }
        }
      }
    },
...
    "/transactor/identities/{id}/eligibility": {
      "get": {
        "summary": "Checks if given id is eligible for free registration",
        "operationId": "Eligibility",
        "parameters": [
          {
            "type": "string",
            "description": "Identity address to register",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Eligibility response",
            "schema": {
              "$ref": "#/definitions/EligibilityResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ErrorMessageDTO": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "x-go-name": "Message",
          "example": "error message"
        }
      },
      "x-go-name": "errorMessage",
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/utils"
    },
...
    "IdentityCreateRequestDTO": {
      "type": "object",
      "title": "IdentityCreateRequest request used for new identity creation.",
      "properties": {
        "passphrase": {
          "type": "string",
          "x-go-name": "Passphrase"
        }
      },
      "x-go-name": "IdentityCreateRequest",
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
    "IdentityCurrentRequestDTO": {
      "type": "object",
      "title": "IdentityCurrentRequest request used for current identity remembering.",
      "properties": {
        "id": {
          "type": "string",
          "x-go-name": "Address"
        },
        "passphrase": {
          "type": "string",
          "x-go-name": "Passphrase"
        }
      },
      "x-go-name": "IdentityCurrentRequest",
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
...
    "IdentityRefDTO": {
      "type": "object",
      "title": "IdentityRefDTO represents unique identity reference.",
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "description": "identity in Ethereum address format",
          "type": "string",
          "x-go-name": "Address",
          "example": "0x0000000000000000000000000000000000000001"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
...
    "ListIdentitiesResponse": {
      "type": "object",
      "title": "ListIdentitiesResponse holds list of identities.",
      "properties": {
        "identities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/IdentityRefDTO"
          },
          "x-go-name": "Identities"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
...
    "ValidationErrorDTO": {
      "type": "object",
      "properties": {
        "errors": {
          "$ref": "#/definitions/FieldErrorMap"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message",
          "example": "error message"
        }
      },
      "x-go-name": "validationErrorMessage",
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/utils"
    },
...
  }
}